+++ src/test/java/kitchenpos/factory/MenuProductFixtureFactory.java
@@ -5,7 +5,7 @@ import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.product.domain.Product;
 
 public class MenuProductFixtureFactory {
-    public static MenuProduct createMenuProduct(Long seq, Menu menu, Product product, long quantity) {
+    public static MenuProduct createMenuProduct(Long seq, Menu menu, Product product, int quantity) {
         return new MenuProduct(seq, menu, product, quantity);
     }
 }
